added an authguard for users that aren't authenticated
