+++ project/web/app/layout.tsx
@@ -1,29 +1,16 @@
 "use client";
 
-
-import { useEffect, useState } from "react";
-import { useRouter, usePathname } from "next/navigation";
-
+import AuthGuard from "./AuthGuard"; 
 import "./globals.css";
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
-  const [, setIsAuthenticated] = useState(false);
-  const router = useRouter();
-  const pathname = usePathname();
-
-  useEffect(() => {
-    const token = localStorage.getItem("authToken");
-    setIsAuthenticated(!!token);
-
-    const authPages = ["/", "/register"];
-    if (!token && !authPages.includes(pathname)) {
-      router.push("/");
-    }
-  }, [pathname, router]);
-  
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body>
+        <AuthGuard>
+          {children}
+        </AuthGuard>
+      </body>
     </html>
   );
 }
